split admin and user api calls from apicalls
